refactor: dto가 도메인을 필드로 갖고 있어서 변경
